back/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
djoser==2.0.3
djangorestframework-simplejwt==4.3.0
uwsgi==2.0.17.1
django-cors-headers==3.1.0
django-cleanup
django-cors-headers
pillow
mysqlclient==1.4.6